+++ src/nitrokey/nk3/secrets_app.py
@@ -9,7 +9,6 @@ import dataclasses
 import hmac
 import logging
 import typing
-from datetime import datetime
 from enum import Enum, IntEnum
 from hashlib import pbkdf2_hmac
 from secrets import token_bytes
@@ -780,7 +779,7 @@ class SecretsApp:
     def verify_pin_raw(self, password: str) -> None:
         structure = [tlv8.Entry(Tag.Password.value, password)]
         self._send_receive(Instruction.VerifyPIN, structure=structure)
-        self.dev._secrets_pin_cache = datetime.now()
+        self.dev.connection.set_secrets_pin_cache()
 
     def get_feature_status_cached(self) -> SelectResponse:
         self._cache_status = self.select() if self._cache_status is None else self._cache_status