+++ src/nitrokey/trussed/__init__.py
@@ -7,7 +7,6 @@
 
 import ctypes
 import sys
-from importlib.util import find_spec
 from typing import List, Optional
 
 from ._base import Model as Model
@@ -17,49 +16,58 @@ from ._bootloader import FirmwareMetadata as FirmwareMetadata
 from ._bootloader import TrussedBootloader as TrussedBootloader
 from ._bootloader import Variant as Variant
 from ._bootloader import parse_firmware_image as parse_firmware_image
-from ._device import App as App
+from ._connection import HAS_CCID_SUPPORT as HAS_CCID_SUPPORT
+from ._connection import App as App
+from ._connection import Transport as Transport
 from ._device import TrussedDevice as TrussedDevice
+from ._exceptions import CcidErrorCode as CcidErrorCode
+from ._exceptions import ConnectionError as ConnectionError
+from ._exceptions import CtapErrorCode as CtapErrorCode
+from ._exceptions import DeviceError as DeviceError
 from ._exceptions import TimeoutException as TimeoutException
 from ._exceptions import TrussedException as TrussedException
 from ._utils import Fido2Certs as Fido2Certs
 from ._utils import Uuid as Uuid
 from ._utils import Version as Version
 
+# module-level constants have no docstrings in Python, so this is documented in
+# docs/api/nitrokey.trussed.rst
+DEFAULT_TRANSPORT = Transport.CTAPHID
 
-def should_default_ccid() -> bool:
-    """Helper function to inform whether CCID should be the default communication protocol
 
-    Some features do not work over CCID, therefore it is only used when CTAPHID is not available, meaning on windows when not an administrator"""
-    if find_spec("smartcard") is None:
-        return False
+def recommended_transport() -> Transport:
+    """Helper function to inform which transport should be used by default.
 
-    if sys.platform != "win32" and sys.platform != "cygwin":
+    Some features do not work over CCID, therefore it is only used when CTAPHID is not available, meaning on windows when not an administrator.
+    See also the :py:const:`DEFAULT_TRANSPORT` constant that defines the default transport if it is not set
+    explicitly."""
+
+    if HAS_CCID_SUPPORT:
         # Linux or MacOS don't need admin to access with CTAPHID
-        return False
+        if sys.platform == "win32" or sys.platform == "cygwin":
+            try:
+                if not ctypes.windll.shell32.IsUserAnAdmin():
+                    return Transport.CCID
+            except Exception:
+                pass
 
-    try:
-        if ctypes.windll.shell32.IsUserAnAdmin():
-            return False
-        else:
-            return True
-    except Exception:
-        return False
+    return Transport.CTAPHID
 
 
 def list(
-    *, use_ccid: bool = False, model: Optional[Model] = None, exclusive: bool = True
+    transport: Transport | None = None, model: Optional[Model] = None, exclusive: bool = True
 ) -> List[TrussedBase]:
     devices: List[TrussedBase] = []
 
     if model is None or model == Model.NK3:
         from nitrokey import nk3
 
-        devices.extend(nk3.list(use_ccid, exclusive))
+        devices.extend(nk3.list(transport=transport, exclusive=exclusive))
 
     if model is None or model == Model.NKPK:
         from nitrokey import nkpk
 
-        devices.extend(nkpk.list(use_ccid, exclusive))
+        devices.extend(nkpk.list(transport=transport, exclusive=exclusive))
 
     return devices